miniforge rshell

+++ install.sh
@@ -47,12 +47,20 @@ rm -rf settings.json
 curl -O https://s3-us-west-2.amazonaws.com/cloudgeniuscode/settings.json
 mv -f settings.json "$HOME/Library/Application Support/Code/User/"
 
-rm -rf $HOME/miniconda ~/miniconda.sh
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O ~/miniconda.sh
-bash ~/miniconda.sh -b -p $HOME/miniconda
-rm -rf ~/miniconda.sh
-source $HOME/miniconda/bin/activate
-conda init zsh
-conda config --set auto_activate_base false
-conda update -n base -c defaults conda -y
-conda update --all -y
+
+if [[ `uname -m` == 'arm64' ]]
+then
+  echo M1 CPU detected - Homebrew preferred prefix /opt/homebrew for Apple Silicon
+  brew install miniforge
+else
+  echo Intel CPU detected - Homebrew preferred prefix /usr/local for macOS Intel
+  rm -rf $HOME/miniconda ~/miniconda.sh
+  wget https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O ~/miniconda.sh
+  bash ~/miniconda.sh -b -p $HOME/miniconda
+  rm -rf ~/miniconda.sh
+  source $HOME/miniconda/bin/activate
+  conda init zsh
+  conda config --set auto_activate_base false
+  conda update -n base -c defaults conda -y
+  conda update --all -y
+fi
